+++ Sources/KeyboardKit/Views/System/SystemKeyboardSpaceButton.swift
@@ -9,54 +9,56 @@
 import SwiftUI
 
 /**
- This view mimics the system space button, which starts with
- displaying the `localeText`, then fading to the `spaceText`.
+ This view mimics a system space button, with custom content,
+ that takes up as much horizontal space as needed.
  
- If you don't provide a `localeText`, this view will use the
- context's current locale. If you don't provide a `spaceText`,
- it will use the localized `KKL10n.space` text.
+ This view will adapt its content to conform to the provided
+ `appearance` and applies keyboard gestures for the provided
+ `actionHandler` and `context`.
  
- Provide empty strings if you don't want to display any text.
- 
- This view creates a `SystemKeyboardSpaceButtonContent` then
- applies a keyboard button style and gestures to it.
+ To mimic a native space button, that starts with displaying
+ a locale text, then fades to the localized name for "space",
+ use `SystemKeyboardSpaceButtonContent` as content.
  */
-public struct SystemKeyboardSpaceButton: View {
+public struct SystemKeyboardSpaceButton<Content: View>: View {
     
     /**
      Create a system keyboard space button.
      
      - Parameters:
-       - localeText: The name of the current locale, if any.
-       - spaceText: The localized name for "space", if any.
-       - actionHandler: The action handler to use.
-       - appearance: The appearance to apply to the button.
+       - actionHandler: The name of the current locale, if any.
+       - appearance: The keyboard appearance to use.
+       - context: The keyboard context to which the button should apply.
+       - content: The content to display inside the button.
      */
     public init(
-        localeText: String,
-        spaceText: String,
         actionHandler: KeyboardActionHandler,
-        appearance: KeyboardAppearance) {
-        self.content = SystemKeyboardSpaceButtonContent(
-            localeText: localeText,
-            spaceText: spaceText)
+        appearance: KeyboardAppearance,
+        context: KeyboardContext,
+        @ViewBuilder content: @escaping ContentBuilder) {
         self.actionHandler = actionHandler
         self.appearance = appearance
+        self.context = context
+        self.content = content()
     }
     
-    private var content: SystemKeyboardSpaceButtonContent
     private let actionHandler: KeyboardActionHandler
     private let appearance: KeyboardAppearance
+    private let context: KeyboardContext
+    private let content: Content
+    
     private var action: KeyboardAction { .space }
     
     @State private var isPressed = false
-    
-    @EnvironmentObject private var context: KeyboardContext
+        
+    /**
+     This typealias represents the configuration action that
+     is used to customize the content of the button.
+     */
+    public typealias ContentBuilder = () -> Content
     
     public var body: some View {
-        content
-            .systemKeyboardButtonStyle(
-                appearance.systemKeyboardButtonStyle(for: action, isPressed: isPressed))
+        button
             .keyboardGestures(
                 for: action,
                 context: context,
@@ -65,14 +67,54 @@ public struct SystemKeyboardSpaceButton: View {
     }
 }
 
+private extension SystemKeyboardSpaceButton {
+    
+    var button: some View {
+        SystemKeyboardButton(
+            content: content.frame(maxWidth: .infinity),
+            style: buttonStyle)
+    }
+    
+    var buttonStyle: SystemKeyboardButtonStyle {
+        appearance.systemKeyboardButtonStyle(
+            for: action,
+            isPressed: isPressed)
+    }
+}
+
 struct SystemKeyboardSpaceButton_Previews: PreviewProvider {
     
-    static var previews: some View {
+    static var spaceText: some View {
+        SystemKeyboardSpaceButtonContent(
+            localeText: KeyboardLocale.english.localizedName,
+            spaceText: KKL10n.space.text(for: .english))
+    }
+    
+    static var spaceView: some View {
+        SystemKeyboardSpaceButtonContent(
+            localeText: KeyboardLocale.spanish.localizedName,
+            spaceView: Text("🇪🇸"))
+    }
+    
+    static func button<Content: View>(for content: Content) -> some View {
         SystemKeyboardSpaceButton(
-            localeText: "foo",
-            spaceText: "bar",
-            actionHandler: .preview,
-            appearance: PreviewKeyboardAppearance())
-            .keyboardPreview()
+            actionHandler: PreviewKeyboardActionHandler(),
+            appearance: .crazy,
+            context: .preview) {
+                content.frame(height: 50)
+            }
+    }
+    
+    static var previews: some View {
+        VStack {
+            Group {
+                button(for: spaceText)
+                button(for: spaceView)
+                button(for: Image.keyboardGlobe)
+            }
+            .padding()
+            .background(Color.red)
+            .cornerRadius(10)
+        }.keyboardPreview()
     }
 }